+++ code/cell.typesystem/src/TypeSystem/TypeClient/TypeClient.fn.validate.ts
