+++ code/cell.typesystem/src/TypeSystem/TypeClient/TypeClient.ts
@@ -1,7 +1,7 @@
 import { t } from '../../common';
 import { util } from '../util';
-import { load } from './fn.load';
-import { typescript } from './fn.typescript';
+import { load } from './TypeClient.fn.load';
+import { typescript } from './TypeClient.fn.typescript';
 
 /**
  * Client that retrieves the type definition of a
